+++ cs-ui-demo/src/ReleaseNotes.md
@@ -8,6 +8,7 @@
   * fixed close button styling when `styleType="light"`
 * CSAutoposition
   * implemented optional chaining of DOMRect objects
+  * fixed autoposition for top and bottom positions
 * CSButton
   * removed icon darken while hovering on `btnStyle="brand"`
   * fixed iconPosition and size props alignment issue